Abstracción del Adaptive Theme para ser utilizado en otras actividades.
